+++ test/cmdlineTests/ethdebug_program_without_ethdebug_in_debug_info/input.sol
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0;
+
+contract C {
+    uint256 value;
+
+    function f(uint256 argument) public {
+        value = argument;
+    }
+}