+++ test/cmdlineTests/standard_output_selection_ethdebug_no_ethdebug_in_debug_info/b.sol
@@ -1,0 +1,4 @@
+//SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0;
+contract A1 { function b(uint x) public pure { assert(x > 0); } }
+contract B2 { function b(uint x) public pure { assert(x > 0); } }